pyproject.toml
[tool.poetry]
name = "myopia"
version = "0.1.0"
description = ""
authors = ["diklios <47837462+diklios5768@users.noreply.github.com>"]
license = "MIT"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.10"
html5lib = "^1.1"
python-dotenv = "^0.20.0"
requests = "^2.27.1"
pycryptodome = "^3.14.1"
cryptography = "^36.0.2"
pandas = "^1.4.2"
openpyxl = "^3.0.9"
PyPDF2 = "^1.27.4"
pillow = "^9.1.0"
pydantic = "^1.9.0"
djantic = "^0.7.0"
Django = "^4.0.4"
django-extensions = "^3.1.5"
#Database
pymysql = "^1.0.2"
django-mysql = "^4.5.0"
django-hashid-field = "^3.3.4"
rules = "^3.3"
django-guardian = "^2.4.0"
pymemcache = "^3.5.2"
pymongo = "^4.1.1"
djongo = "^1.3.6"
#API
#django-ninja="*"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
#drf-yasg = "*"
drf-spectacular = "^0.22.0"
django-simpleui = "^2022.4.9"
django-cors-headers = "^3.11.0"
django-sslserver = "^0.22"
#Server
gevent = "^21.12.0"
waitress = { version = "^2.1.1", platform = "win32" }
gunicorn = { version = "^20.1.0", markers = "sys_platform != 'win32'" }



[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pylint = "^2.13.9"
pretty-errors = "^1.2.25"
pytest = "^7.1.1"
#jinja2 = "*"
#django-jinja = "*"
django-sslserver = "^0.22"
#pdfplumber = "^0.6.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
